[WIP] ATA: bump ACS version
This prepares us for the use of TRUSTED SEND/RECV commands to be used
for DMTF SPDM support.

Note: We are lying about complete compliance here, there are features of
the ATA8 set that aren't yet supported. But for TRUSTED SEND/RECV
commands to work, ATA8 minimum is required from the linux kernel side.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>